rename deploy -> sharing
